Moved patientRecords state back to App.
